Conv+l2norm fuse pass test (manually check, dst_data_t=s32)

+++ tests/gtests/test_convolution_forward_u8s8s32.cpp
@@ -21,10 +21,10 @@
 #include "test_convolution_forward_common.hpp"
 namespace mkldnn {
 
-//using convolution_test = convolution_forward_test<uint8_t, int8_t,
-//                                                int32_t, int32_t>;
 using convolution_test = convolution_forward_test<uint8_t, int8_t,
-                                                int32_t, float>;
+                                                int32_t, int32_t>;
+//using convolution_test = convolution_forward_test<uint8_t, int8_t,
+//                                                int32_t, float>;
 
 TEST_P(convolution_test, TestConvolution)
 {
